requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
# google-auth-httplib2==0.2.1
google-auth-httplib2==0.2.0

google-api-python-client==2.80.0
pydantic==1.10.0
gspread==5.0.0
oauth2client==4.1.3
openai==1.3.5
